chore: delete home folder, rename frontend to frontend-old and frontend-new to frontend
