+++ src/app/loaders/express/index.ts
@@ -11,8 +11,8 @@ import { AuthRouter } from '../../modules/auth/auth.routes'
 import { ExamplesRouter } from '../../modules/examples/examples.routes'
 import { AdminFormsRouter } from '../../modules/form/admin-form/admin-form.routes'
 import { PublicFormRouter } from '../../modules/form/public-form/public-form.routes'
-import { FrontendRouter as OldFrontendRouter } from '../../modules/frontend/frontend.routes'
-import { FrontendRouter } from '../../modules/frontend-new/frontend.routes'
+import { FrontendRouter } from '../../modules/frontend/frontend.routes'
+import { FrontendRouter as OldFrontendRouter } from '../../modules/frontend-old/frontend.routes'
 import { MYINFO_ROUTER_PREFIX } from '../../modules/myinfo/myinfo.constants'
 import { MyInfoRouter } from '../../modules/myinfo/myinfo.routes'
 import { SgidRouter } from '../../modules/sgid/sgid.routes'
