--- src/app/modules/home/home.routes.ts
@@ -1,7 +1,0 @@
-import { Router } from 'express'
-
-import * as HomeController from './home.controller'
-
-export const HomeRouter = Router()
-
-HomeRouter.get('/', HomeController.home)
